🔧 ▸ Fixing dev (check desc for solution)
The blueprints was not able to get the c++ files because it has not compile, but if you tryied compiling you will get errors (and fucked).

The solution was to delete the Intermediate folder, and to find that the .cpp enum files has hard coded path file in it. So I deleted them.